[blocks] organize helpers into mixins
