--- starport/ui/src/mixins/blocks.js
@@ -1,7 +1,0 @@
-export default {
-  data() {
-    return {}
-  },
-  methods: {
-  }
-}
